refactor(B-INT-6): decompose BicameralRoute.ts under the Section-4 razor
490 -> 34 LoC orchestrator. Extract bicameralRouteShared.ts (deps + governToolCall
+ helpers), bicameralLifecycleRoutes.ts, bicameralDecisionRoutes.ts; repoint
bicameralToolRoutes.ts at the shared core (breaks the prior import cycle). Public
surface preserved via re-export. Zero behavioral change: 195 mocha + 7 Playwright
verbatim. META_LEDGER #410.
